+++ Civi/DataProcessor/FieldOutputHandler/MarkupFieldOutputHandler.php
@@ -1,0 +1,32 @@
+<?php
+/**
+ * @author Jaap Jansma <jaap.jansma@civicoop.org>
+ * @license AGPL-3.0
+ */
+
+namespace Civi\DataProcessor\FieldOutputHandler;
+
+use Civi\DataProcessor\Exception\DataSourceNotFoundException;
+use Civi\DataProcessor\Exception\FieldNotFoundException;
+use CRM_Dataprocessor_ExtensionUtil as E;
+use Civi\DataProcessor\Source\SourceInterface;
+use Civi\DataProcessor\DataSpecification\FieldSpecification;
+
+class MarkupFieldOutputHandler extends AbstractSimpleFieldOutputHandler {
+
+  /**
+   * Returns the formatted value
+   *
+   * @param $rawRecord
+   * @param $formattedRecord
+   *
+   * @return \Civi\DataProcessor\FieldOutputHandler\FieldOutput
+   */
+  public function formatField($rawRecord, $formattedRecord) {
+    $formattedValue = $rawRecord[$this->inputFieldSpec->alias];
+    $output = new HTMLFieldOutput($rawRecord[$this->inputFieldSpec->alias]);
+    $output->setHtmlOutput($formattedValue);
+    return $output;
+  }
+
+}